test(process): isolate foreground retention regression
